The addition mod procedure and creation is now implemented!

+++ modmanager/ModPackage.cs
@@ -135,23 +135,28 @@ namespace modmanager
 
 			for(int i = 0; i < ModCount; i++)
 			{
-				backup_path = Path.Combine(p.BackupRoot, Mods[i].BackupFile);
-
-				if(!File.Exists(backup_path))
+				//Backups only make sense in the context of replacement mods
+				if(Mods[i].ModType == Mod.Type.Replacement)
 				{
-					string file_path = Path.Combine(p.GamePath, Mods[i].BackupFile);
-					string backup_dir_path = Utils.GetLastDirectory(backup_path);
+					backup_path = Path.Combine(p.BackupRoot, Mods[i].BackupFile);
 
-					try
+					if (!File.Exists(backup_path))
 					{
-						Directory.CreateDirectory(backup_dir_path);
-						File.Copy(file_path, backup_path);
-					}
-					catch(Exception e)
-					{
-						System.Windows.Forms.MessageBox.Show("Error creating backup!\n\nFROM: " + file_path + "\nTO:" + backup_path + "\n\n" + e.Message);
+						string file_path = Path.Combine(p.GamePath, Mods[i].BackupFile);
+						string backup_dir_path = Utils.GetLastDirectory(backup_path);
+
+						try
+						{
+							Directory.CreateDirectory(backup_dir_path);
+							File.Copy(file_path, backup_path);
+						}
+						catch (Exception e)
+						{
+							System.Windows.Forms.MessageBox.Show("Error creating backup!\n\nFROM: " + file_path + "\nTO:" + backup_path + "\n\n" + e.Message);
+						}
 					}
 				}
+				
 			}
 		}
 	}
